Add files via upload
This code takes a list of violent phrases and a set of messages, cleans both datasets, and then identifies occurrences of violent phrases within the messages, outputting results to Excel files.  The phrase matching is designed to be relatively flexible, accounting for minor variations in the wording of the phrases.